Abstracted Forge and NeoForge installers into unified Forgelike classes
