+++ app_pojavlauncher/src/main/java/net/kdt/pojavlaunch/modloaders/VersionResolver.java
@@ -1,0 +1,11 @@
+package net.kdt.pojavlaunch.modloaders;
+
+/**
+ * Resolves the loader version from a minecraft version and a loader version
+ *
+ * Used in {@link ForgelikeUtils}
+ */
+@FunctionalInterface
+public interface VersionResolver {
+    String resolveVersion(String minecraftVersion, String modloaderVersion);
+}
